+++ src/main/kotlin/me/ffl/intellijDirectoryTests/Highlighting.kt
@@ -4,8 +4,8 @@ import com.intellij.openapi.application.runWriteAction
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.openapi.vfs.writeText
 import com.intellij.testFramework.VfsTestUtil
-import com.intellij.util.io.isDirectory
 import java.nio.file.Path
+import kotlin.io.path.isDirectory
 import kotlin.io.path.listDirectoryEntries
 import kotlin.io.path.readText